requirements.txt
# Python Dependencies for Credit Risk Analysis Project

# Data Processing
pandas>=1.5.0
numpy>=1.23.0

# Database Connectivity
pymysql>=1.0.2

# Environment Variables
python-dotenv>=1.0.0

# Utilities
tabulate>=0.9.0

# API Server (FastAPI)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Automated Reporting
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Google Sheets Integration (Optional)
gspread>=5.11.0
google-auth>=2.23.0

# Optional: Kaggle API (for Home Credit dataset)
kaggle>=1.5.12

# Optional: Data Visualization (if using Python instead of Tableau)
plotly>=5.0.0
dash>=2.0.0

# Optional: SQLAlchemy (alternative ORM)
# sqlalchemy>=2.0.0
